Move CancellationToken to threading directory and update includes
